small tweak to remove warning on some compilers
git-svn-id: svn+ssh://nextgen.cshl.edu/usr/local/svnroot/phast/trunk@6261 3e584e52-61a1-4911-b546-322c73959ec4

+++ include/lists.h
@@ -72,7 +72,7 @@ void *srealloc(void *ptr, size_t size);
 
 static PHAST_INLINE
 void lst_arr_set(List *l, int i, void *o) {
-  if (l->elementsz <= sizeof(void*))
+  if (((unsigned int)l->elementsz) <= sizeof(void*))
     l->array[i] = *((void**)o);	/* ?? */
   else
     memcpy(&l->array[i * l->step], o, l->elementsz);
